Primera refactorización del codigo
