+++ modules/info.py
@@ -1,0 +1,47 @@
+from tkinter import messagebox
+
+# Functions
+
+def arg1985():
+    messagebox.showinfo(title="ARGENTINA 1985",
+        message="Durante la década de 1980, un grupo de abogados investiga y lleva a juicio a los responsables de la dictadura cívico-militar argentina.                                                                                  GENERO:Drama                                                               DURACION: 2h 20m                                                               DIRECTOR: Santiago Mitre")
+
+def avatar():
+    messagebox.showinfo(title="AVATAR 2",
+        message="Jake Sully y Ney'tiri han formado una familia y hacen todo lo posible por permanecer juntos. Sin embargo, deben abandonar su hogar y explorar las regiones de Pandora cuando una antigua amenaza reaparece.                                    GENERO:Ciencia Ficcion/Aventura                                  DURACION: 166min                                                        DIRECTOR: James Cameron ")
+
+def dontworrydarling():
+    messagebox.showinfo(title="DON'T WORRY DARLING", 
+    message= "Alice y Jack, todo parece ir muy bien en la vida de ambos, pero en poco tiempo, Alice descubre un proyecto de la empresa en donde trabaja Jack y él está involucrado: el misterioso Proyecto Victoria                                                                 GENERO:Suspenso                                                               DURACION: 2h 3m                                                        DIRECTORA:Olivia Wilde")
+
+def lahuerfana():
+    messagebox.showinfo(title="LA HUERFANA: EL ORIGEN", 
+    message= "Lena se escapa del psiquiátrico y viaja a EE. UU. haciéndose pasar por la hija desaparecida de una familia adinerada. Pero su nueva vida como Esther no será como ella esperaba.            GENERO:Terror/Suspenso                                                           DURACION: 1h 39m                                                                 DIRECTOR: William Brent Bell")
+
+
+def sonrie():
+    messagebox.showinfo(title="SONRIE", 
+    message="Tras la presencia de un dramático incidente sufrido por un paciente, la Dra. Cotter empieza a experimentar hechos aterradores sin explicación aparente.                                      GENERO:Terror Psicológico                                                        DURACION:1h 55m                                                              DIRECTOR: Parker Finn")
+
+def blackadam():
+    messagebox.showinfo(title="BLACK ADAM", 
+    message= "Black Adam es una próxima película estadounidense de superhéroes basada en el personaje homónimo de DC Comics y pretende ser una derivación de SHAZAM!.                                 GENERO: Accion/Fantasia                                                                DURACION:2h 4m                                                     DIRECTOR:Jaume Collet-Serra")
+
+
+def thor():
+    messagebox.showinfo(title="THOR: AMOR Y TRUENO", 
+    message="Thor está buscando la paz interior, pero la irrupción de Gorr, le obliga a volver al combate. Thor recluta a su gente para que lo ayuden a evitar la extinción de los dioses.            GENERO:Accion/Aventura                                                               DURACION: 1h 59m                                               DIRECTOR:Taika Waititi")
+
+def losminions():
+    messagebox.showinfo(title="MINIONS: NACE UN VILLANO", 
+    message="En los años 70, Gru crece siendo un gran admirador de Los salvajes seis, un supergrupo de villanos. Para demostrarles que puede ser malvado, diseña un plan con la esperanza de formar parte de la banda.                                                      GENERO:Comedia                                                                          DURACION: 1h 30m                                                                 DIRECTOR: Kyle Balda")
+
+def topgun():
+    messagebox.showinfo(title="TOP GUN: MAVERICK", 
+    message="Maverick, quien lleva 30 años de servicio, es ahora instructor de pilotos militares. Una última misión, Es obligado a enfrentar las heridas abiertas del pasado y sus temores más profundos.                                                                    GENERO:Accion                                                                  DURACION:2h 11m                                                          DIRECTOR:Joseph Kosinski")
+
+def blackpanther():
+    messagebox.showinfo(title="BLACK PANTHER 2: WAKANDA FOREVER", 
+    message="Una secuela que seguirá explorando el incomparable mundo de Wakanda y todos los ricos y variados personajes presentados en la película de 2018.                                     GENERO:Accion                                                                DURACION:2h 41m                                                          DIRECTOR:Ryan Coogler")
+
+#---------------------------------------------
